requirements.txt
django==4.2.29
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==11.3.0
python-dotenv==1.2.1
